requirements.txt
fastapi
uvicorn
pydantic
yt-dlp
faster-whisper
fastembed
pinecone
python-dotenv
